Moved VennDiagramRegion to halo_simulator.py and changed the signautre of simulated_venn_diagram_reach_by_spend.
